perutimber: Catalogue of the Timber Forest Species of the Peruvian Amazon

Access the data of the 'Catalogue of the Timber Forest Species of the Peruvian Amazon' Vásquez Martínez, R., & Rojas Gonzáles, R.D.P.(2022)doi:10.21704/rfp.v37i3.1956 .

See also

Author

Maintainer: Paul E. Santos Andrade paulefrens@gmail.com (ORCID)