List of Plant Species Name and information from the Catalogue of the timber forest species of the Amazon and the Peruvian Yunga.

The 'perutimber_dat' contains information from the Catalogue of the timber forest species of the Amazon and the Peruvian Yunga.

Usage

perutimber_data

Format

A tibble with the following columns:

accepted_name_author

A character vector. The corresponding author name of the accepted_name, staying empty if the taxonomic_status is "Synonym" or "No opinion".

accepted_family

A character vector. The corresponding family name of the accepted_name.

common_name

A character vector. Listed all common names for each species.

accepted_name_rank

A character vector. Information about infraspecific status.

habit

A character vector.

plant_height_m

A character vector. The plant height range in meters.

regions

A character vector.

elevation

A character vector.

uses

A character vector.

Examples

data(perutimber_data)
str(perutimber_data)
#> 'data.frame':	1339 obs. of  10 variables:
#>  $ accepted_name       : chr  "Abarema acreana" "Abarema adenophora" "Abarema auriculata" "Abarema jupunba" ...
#>  $ accepted_name_author: chr  "(J.F. Macbr.) L. Rico" "(Ducke) Barneby & J.W. Grimes" "(Benth.) Barneby & J.W. Grimes" "(Willd.) Britton & Killip" ...
#>  $ accepted_family     : chr  "Fabaceae" "Fabaceae" "Fabaceae" "Fabaceae" ...
#>  $ common_name         : chr  "Pashaco blanco" "No registra" "No registra" "Chunkámachari, Kuúp, Pashaquilla amarilla" ...
#>  $ accepted_name_rank  : chr  "species" "species" "species" "species" ...
#>  $ habit               : chr  "Árbol" "Árbol" "Árbol" "Árbol" ...
#>  $ plant_height_m      : chr  "30-35(-45) m." "(10-) 20-30 m." "25-35(-40) m." "20-30 m." ...
#>  $ regions             : chr  "LO." "AM, LO, PA." "LO." "AM, JU, LO, MD, PA, PU, SM." ...
#>  $ elevation_m         : chr  "100-500" "100-500" "100-500" "100-1000(-1500)" ...
#>  $ uses                : chr  "madera aserrada" "con aptitud maderable" "con aptitud maderable" "madera aserrada" ...