Search species name present in the Catalogue of the timber forest species of the Amazon and the Peruvian Yunga

This function takes a species list and tries to match a name in the Catalogue of the timber forest species of the Amazon and the Peruvian Yunga, checking if the name is listed in tha dataset.

Usage

search_perutimber(splist, max_distance = 0.2)

Arguments

splist

A character vector specifying the input taxon, each element including genus and specific epithet and, potentially, infraspecific rank, infraspecific name and author name. Only valid characters are allowed (see base::validEnc).

max_distance

match when comparing the submitted name with the closest name matches in the species listed in the "Catalogue of the timber forest species of the Amazon and the Peruvian Yunga". The distance used is a generalized Levenshtein distance that indicates the total number of insertions, deletions, and substitutions allowed to match the two names. It can be expressed as an integer or as the fraction of the binomial name. For example, a name with length 10, and a max_distance = 0.1, allow only one change (insertion, deletion, or substitution). A max_distance = 2, allows two changes.

Value

A character vector that can have three different output values. The first value, "Present," indicates whether the species name is fully matched with the names listed in the 'Catalogue of the timber forest species of the Amazon and the Peruvian Yunga'. The second value, "P_updated_name," provides a fuzzy matching of species names. If a species name is not listed in the catalogue, the third value returned will be an empty string.

Examples

# Search for multiple species vector
splist <- c("Euterpe precatoria var. precatorio",
            "Welfia alfredi",
            "Hibiscus abelmoschus var. betulifolius")
get_perutimber_data(splist)
#>                      names_submitted                      accepted_name
#> 1 Euterpe precatoria var. precatorio Euterpe precatoria var. precatoria
#> 2                     Welfia alfredi                    Welfia alfredii
#>    accepted_name_author accepted_family         common_name accepted_name_rank
#> 1                             Arecaceae Huasaí, Saké, Tsáke            variety
#> 2 A.J. Hend. & Villalba       Arecaceae              Camona            species
#>           habit plant_height_m             regions     elevation_m
#> 1 Palma arbórea   8-18(-25) m. AM, LO, MD, PA, SM. 100-1000(-2000)
#> 2 Palma arbórea  10-20(-25) m.                 PA.       1400-1700
#>                    uses
#> 1       madera aserrada
#> 2 con aptitud maderable

# Search for multiple species data.frame
# base
df_splist <- data.frame(splist = splist)
df_splist$peutimber <- search_perutimber(df_splist$splist)