List of Plant Species Name accordingly with the Catalogue of the timber forest species of the Amazon and the Peruvian Yunga.

The 'tab_perutimber' contains records belonging to all the timber forest species the Amazon and the Peruvian Yunga.

Usage

tab_perutimber

Format

A tibble with the following columns:

input_genus

A character vector. The input genus of the corresponding species name listed in CTSAPY.

input_epitheton

A character vector. The specific epithet of the corresponding species name listed in CTSAPY.

rank

A character vector. The taxonomic rank: "species","subspecies", "variety", of the corresponding species name listed in CTSAPY.

input_subspecies_epitheton

A character vector. If the indicated rank is below species, the subspecies epithet input of the corresponding species name listed in CTSAPY.

taxonomic_status

A character vector. description if a taxon is classified as ‘accepted’, ‘synonym’, ‘no opinion’. According to the Taxonomic Name Resolution Service - TNRS.

accepted_name

A character vector. The list of the accepted plant taxa names according to the Taxonomic Name Resolution Service - TNRS.

accepted_family

A character vector. The corresponding family name of the accepted_name.

accepted_name_author

A character vector. The corresponding author name of the accepted_name, staying empty if the taxonomic_status is "Synonym" or "No opinion".

tnrs_name_id

A character vector. The accepted_name_id according to the Taxonomic Name Resolution Service - TNRS output.

accepted_name_url

A character vector. The accepted_name information url.

source

A character vector. The accepted_name information source.

References

Catalogue of the timber forest species of the Amazon and the Peruvian Yunga. Boyle, B. L., Matasci, N., Mozzherin, D., Rees, T., Barbosa, G. C., Kumar Sajja, R., & Enquist, B. J. (2021). Taxonomic Name Resolution Service, version 5.0. In Botanical Information and Ecology Network. https://tnrs.biendata.org/ , Accessed Mar 21, 2023.

Examples

data(tab_perutimber)
str(tab_perutimber)
#> 'data.frame':	1343 obs. of  12 variables:
#>  $ id_cat                    : chr  "1LIPA" "2SASP" "3ALVE" "4ANGI" ...
#>  $ input_genus               : chr  "Lindackeria" "Saurauia" "Alzatea" "Anacardium" ...
#>  $ input_epitheton           : chr  "paludosa" "spectabilis" "verticillata" "giganteum" ...
#>  $ rank                      : chr  "species" "species" "subspecies" "species" ...
#>  $ input_subspecies_epitheton: chr  "null" "null" "verticillata" "null" ...
#>  $ taxonomic_status          : chr  "Accepted" "Accepted" "Accepted" "Accepted" ...
#>  $ accepted_name             : chr  "Lindackeria paludosa" "Saurauia spectabilis" "Alzatea verticillata subsp. verticillata" "Anacardium giganteum" ...
#>  $ accepted_family           : chr  "Achariaceae" "Actinidiaceae" "Alzateaceae" "Anacardiaceae" ...
#>  $ accepted_name_author      : chr  "(Benth.) Gilg" "Hook." "null" "W. Hancock ex Engl." ...
#>  $ tnrs_name_id              : chr  "255690" "4951" "9073" "13175" ...
#>  $ accepted_name_url         : chr  "http://www.tropicos.org/Name/13200675" "http://www.tropicos.org/Name/500185" "http://www.tropicos.org/Name/1000003" "http://www.tropicos.org/Name/1300075" ...
#>  $ source                    : chr  "tropicos" "tropicos" "tropicos" "tropicos" ...